Add `router.go` and `backend.go`: Extend proxy framework with routing and backend configurations
